refactor: removed unused classes and moved settings types to core
