+++ sdk/src/core/settings/walletConnect/HWCSettings.ts
