+++ sdk/src/domain/context/kyc/Terminal3.ts
