+++ sdk/src/domain/context/shared/error/InvalidBytes3.ts
