+++ sdk/src/port/in/request/validation/Validation.ts
@@ -225,7 +225,7 @@ import { EVM_ZERO_ADDRESS } from '../../../../core/Constants.js';
 import { InvalidEvmAddress } from '../../../../domain/context/contract/error/InvalidEvmAddress.js';
 import { InvalidFormatHederaIdOrEvmAddress } from '../error/InvalidFormatHederaIdOrEvmAddress.js';
 import { InvalidBytes32 } from '../../../../domain/context/security/error/InvalidBytes32.js';
-import { InvalidBytes3 } from '../../../../domain/context/bond/error/InvalidBytes3.js';
+import { InvalidBytes3 } from '../../../../domain/context/shared/error/InvalidBytes3.js';
 import { HEDERA_FORMAT_ID_REGEX } from '../../../../domain/context/shared/HederaId.js';
 import { InvalidBytes } from '../../../../domain/context/shared/error/InvalidBytes.js';
 import { InvalidBase64 } from '../../../../domain/context/shared/error/InvalidBase64.js';
